+++ packages/webapp-libs/webapp-core/src/components/forms/form/formControl/index.ts
@@ -1,0 +1,1 @@
+export * from './formControl.component';
