+++ packages/webapp-libs/webapp-core/src/components/forms/form/formDescription/formDescription.component.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+import { useFormField } from '../formField';
+import { cn } from '@sb/webapp-core/lib/utils';
+
+const FormDescription = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<HTMLParagraphElement>>(
+  ({ className, ...props }, ref) => {
+    const { formDescriptionId } = useFormField();
+
+    return <p ref={ref} id={formDescriptionId} className={cn('text-sm text-muted-foreground', className)} {...props} />;
+  }
+);
+FormDescription.displayName = 'FormDescription';
+
+export { FormDescription };
