+++ packages/webapp-libs/webapp-core/src/components/forms/form/formDescription/index.ts
@@ -1,0 +1,1 @@
+export * from './formDescription.component';
