+++ packages/webapp-libs/webapp-core/src/components/forms/form/formField/index.ts
@@ -1,0 +1,1 @@
+export * from './formField.component';
