+++ packages/webapp-libs/webapp-core/src/components/forms/form/formItem/formItem.component.tsx
@@ -1,0 +1,23 @@
+import * as React from 'react';
+import { cn } from '@sb/webapp-core/lib/utils';
+
+type FormItemContextValue = {
+  id: string;
+};
+
+const FormItemContext = React.createContext<FormItemContextValue>({} as FormItemContextValue);
+
+const FormItem = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
+  ({ className, ...props }, ref) => {
+    const id = React.useId();
+
+    return (
+      <FormItemContext.Provider value={{ id }}>
+        <div ref={ref} className={cn('space-y-2', className)} {...props} />
+      </FormItemContext.Provider>
+    );
+  }
+);
+FormItem.displayName = 'FormItem';
+
+export { FormItem, FormItemContext };
