+++ packages/webapp-libs/webapp-core/src/components/forms/form/formItem/index.ts
@@ -1,0 +1,1 @@
+export * from './formItem.component';
