+++ packages/webapp-libs/webapp-core/src/components/forms/form/formLabel/index.ts
@@ -1,0 +1,1 @@
+export * from './formLabel.component';
