+++ packages/webapp-libs/webapp-core/src/components/forms/form/formMessage/formMessage.component.tsx
@@ -1,0 +1,24 @@
+import * as React from 'react';
+
+import { useFormField } from '../formField';
+import { cn } from '@sb/webapp-core/lib/utils';
+
+const FormMessage = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<HTMLParagraphElement>>(
+  ({ className, children, ...props }, ref) => {
+    const { error, formMessageId } = useFormField();
+    const body = error ? String(error?.message) : children;
+
+    if (!body) {
+      return null;
+    }
+
+    return (
+      <p ref={ref} id={formMessageId} className={cn('text-sm font-medium text-destructive', className)} {...props}>
+        {body}
+      </p>
+    );
+  }
+);
+FormMessage.displayName = 'FormMessage';
+
+export { FormMessage };
