+++ packages/webapp-libs/webapp-core/src/components/forms/form/formMessage/index.ts
@@ -1,0 +1,1 @@
+export * from './formMessage.component';
