+++ packages/webapp-libs/webapp-core/src/components/forms/form/index.ts
@@ -1,1 +1,7 @@
 export * from './form.component';
+export * from './formControl';
+export * from './formDescription';
+export * from './formField';
+export * from './formItem';
+export * from './formLabel';
+export * from './formMessage';
